requirements.txt
coverage==6.3.1
geopy==2.2.0
python-telegram-bot==13.11
pytest==7.0.1
pytest-cov==3.0.0
